requirements.txt
openai
python-dotenv
flask>=3.0.0
